feat: 상품 상세 컴포넌트 개발 후 상품 상세 페이지에서 분리
