+++ src/api/product.ts
@@ -1,4 +1,5 @@
 import { API_URL, PRODUCT_LIST_PAGE_LIMIT } from '@/api/constants';
+import { ProductType } from '@/domain/product';
 import axios from 'axios';
 
 const productAPI = axios.create({
@@ -17,7 +18,7 @@ export const getProductList = async (page): Promise<any> => {
   return { productList: response.data, totalProductCount: response.headers['x-total-count'] };
 };
 
-export const getProduct = async (id): Promise<any> => {
+export const getProduct = async (id): Promise<{ product: ProductType }> => {
   const response = await productAPI.get(`/${id}`, {});
   if (response.statusText !== 'OK') {
     throw Error('서버 오류!');
