+++ src/pages/ProductDetail/ProductDetail.tsx
@@ -1,23 +1,20 @@
 import { getProduct } from '@/api/product';
-import Button from '@/components/common/Button/Button';
-import Image from '@/components/common/Image/Image';
+
 import Loading from '@/components/common/Loading/Loading';
-import Modal from '@/components/common/Modal/Modal';
-import CartAdd from '@/components/product/CartAdd/CartAdd';
+import ProductDetailCard from '@/components/product/ProductDetailCard/ProductDetailCard';
+import { ProductType } from '@/domain/product';
+
 import { useThunkFetch } from '@/hooks/useFecth';
-import { useModal } from '@/hooks/useModal';
-import useResponsive from '@/hooks/useResponsive';
+
 import { fetchGetCartAsync } from '@/store/cart/action';
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router';
 import ErrorContainer from '../../components/common/ErrorContainer/ErrorContainer';
 import PageTemplate from '../../components/common/PageTemplate/PageTemplate';
+
 import * as Styled from './ProductDetail.style';
 
 function ProductDetail() {
-  const responsive = useResponsive();
-  const { isShowModal, openModal, closeModal } = useModal();
-
   const { productId } = useParams();
 
   const { isLoading: isCartLoading, cartList } = useThunkFetch({
@@ -26,7 +23,10 @@ function ProductDetail() {
     deps: [],
   });
 
-  const [{ isLoading, product }, setProduct] = useState({
+  const [{ isLoading, product }, setProduct] = useState<{
+    isLoading: boolean;
+    product: ProductType | null;
+  }>({
     isLoading: false,
     product: null,
   });
@@ -62,43 +62,13 @@ function ProductDetail() {
     );
   }
 
-  const onClickCartAddBUtton = () => {
-    openModal();
-  };
-
-  const isShowCartButton = !cartList.find(cart => cart.id === (product as any)?.id);
-
   return (
     <PageTemplate>
       <Styled.Container>
-        <Styled.Title>상품 상세</Styled.Title>
-
-        <Styled.InformationWrapper>
-          <Image
-            src={(product as any).imageURL}
-            alt=""
-            width={responsive === 'desktop' ? '400px' : '250px'}
-          />
-          <Styled.Name>{(product as any).name}</Styled.Name>
-          <Styled.Price>
-            <span>금액 </span>
-            <span>{(product as any).price}원</span>
-          </Styled.Price>
-        </Styled.InformationWrapper>
-
-        {isShowCartButton && (
-          <Styled.ButtonWrapper onClick={onClickCartAddBUtton}>
-            <Button width="100%" padding="20px">
-              장바구니
-            </Button>
-          </Styled.ButtonWrapper>
-        )}
-
-        {isShowModal && (
-          <Modal closeModal={closeModal}>
-            <CartAdd product={product} closeModal={closeModal} />
-          </Modal>
-        )}
+        <ProductDetailCard
+          product={product}
+          isShowCartAddButton={!cartList.find(cart => cart.id === (product as any)?.id)}
+        />
       </Styled.Container>
     </PageTemplate>
   );
